fix bad interaction with shiv, vulcan, and pipx
When vulcan is installed with pipx, the `shiv` executable is not placed
on your path because pipx only installs the direct console_scripts, not
the ones of dependent libraries. This makes sense from the point of view
of avoid conflict, but vulcan depends on that shiv command. Luckily,
this is a way to access it that is actually much more reliable because
it ensures that not only do we always find shiv, we always find the
_correct_ shiv.